refactor(ssaapi): split compile-unit helpers and tighten default observability
Move batching, pprof/GC, and logging out of ssa_compile_fs.go, fix dbcache
Flush to drain buffered saves, restore ssalog for SSA cache paths, and make
adaptive GC plus split-compile debug output opt-in so normal compiles stay quiet.
