+++ common/yak/ssaapi/compile_units_test.go
@@ -107,15 +107,22 @@ func TestCompileUnitWriterCacheEnabledForMultiBatchOrLargeProject(t *testing.T) 
 func TestSSACompileAdaptiveGCPolicyDefaultsAndOverrides(t *testing.T) {
 	t.Setenv("GOGC", "")
 	t.Setenv("GOMEMLIMIT", "")
+	t.Setenv(ssaCompileAdaptiveGCEnv, "")
 	t.Setenv(ssaCompileGOGCEnv, "")
 	t.Setenv(ssaCompileMemLimitEnv, "")
 
 	gcPercent, setGC := ssaCompileGCPercent()
 	memLimit, setMemLimit := ssaCompileMemoryLimit()
-	require.True(t, setGC)
+	require.False(t, setGC)
 	require.Equal(t, defaultSSACompileGOGC, gcPercent)
-	require.True(t, setMemLimit)
+	require.False(t, setMemLimit)
 	require.Equal(t, int64(defaultSSACompileMemLimit), memLimit)
+	require.Nil(t, startSSACompileAdaptiveGC(nil))
+
+	t.Setenv(ssaCompileAdaptiveGCEnv, "1")
+	stop := startSSACompileAdaptiveGC(nil)
+	require.NotNil(t, stop)
+	stop()
 
 	t.Setenv(ssaCompileGOGCEnv, "120")
 	t.Setenv(ssaCompileMemLimitEnv, "512MiB")
